Provide a XPK decoder in C
There's a lot going on here, so let me explain. In prinicple, it's a fairly
straightforward reimplementation of what the Ruby xpkread tool is doing, but
I ended up inventing names for the different XPK instructions, and I had to
augment the XPK and XPKEntry structures to expose more information to the
new XPKDecoder struct.

The decoder looks intimidating, but if you look closely, it's just an
enormous switch() statement. And it's going to get bigger...  there are
still bugs, looks like there's a bunch of opcodes we haven't figured out,
yet (0x1 and 0x9, it seems).