XPK: Recognize the 0x20 and 0xA0 instructions
Take 0x20 for example. It says "XSKIP 0" or "advance cursor in X direction
by zero pixels," which doesn't make sense, so 0x20 must mean something else:
specifically, read the next 2 bytes, and this 16-bit value represents the
count by which to skip in the X direction.

A similar story is true for the 0xA0 instruction, but a newline is implied,
too as usual with SETXNL.