Import getopt(3) from alex85k/wingetopt, use it iff building for Windows
Windows' C runtime does not come with an implementation of getopt(3), so we
bring in this one when required. It's originally based on NetBSD's version,
but with some Windows-specific flavor by the mingw folks.

The major caveat here is the following: the CMake now suggests the only time
where one wouldn't have getopt(3) is if you're building for Windows. It's
probably healthier to directly check whether getopt() is available, but
getopt() is definitely available on all the other OSes I care about (OS X,
*BSD, Linux), so that's fine by me. And the main theme of the day is to get
Windows support working, anyway.