Introducing LP{Log,Warn,Debug}, new logging functions which avoid warnx(3)
This is a complete overhaul of stdout/stderr logging.

The primary goal here is to avoid warnx(3) and including err.h altogether,
which are BSD extensions to C.

The main motivation behind this change is for Windows' benefit, but actually
the clearer semantics with "Log", "Warn" and "Debug" is nice in general.