PRX: There might be "dummy" entries at the *end* of the TOC, too
There's this one element of PRX archives which we have been skipping. It
tells you how many members there *really* are, including dummy entries,
which may be at the beginning *and/or* the end of the table of contents.
Take this extra information into account, to avoid the possibility of
accidentally extracting these empty archive members.

Fixes #28