Allow creating a CLU struct from any region of memory. Introduce test programs.
It kinda sucks to have to introduce the test programs framework in _this_
commit, but at least we can demonstrate that CLU_NewFromData() actually
does the right thing.

CLU_NewFromFile() is rewritten to be in terms of CLU_NewFromData(). There's
probably a smarter way to handle memory consumption in that case, but
reading an entire CLU into memory takes up only 2056 bytes, which is not
*that* bad, really.

Fixes #40