XPK: Add support for "tiled mode"
Opting into "tiled mode" signals your intention to load all 64 entries in an
XPK (80x60 each) in order to assemble a single 640x480 image instead. This
is used to create full-screen background images, such as while in the Hidden
Hallway or during a choice screen.

As far as we can tell, the XPK data itself does not explicitly state whether
the image is to be interpreted in this tiled mode, or as individual frames
in a single animation, or anything like that. I guess the user needs to
declare up-front whether the XPK ought to be decoded in this manner.