Make an API which can read PRD/PRS pairs *AND* PRX archives.
In this changeset, we totally reimagine how to deal with Presage's multiple
archive formats, while taking into consideration the fact that they're
*mostly* similar with one another.

While here, we take the opportunity to reduce memory usage, in a style
similar to the MMFW work that was done recently. In short, reading a PRX (or
PRD/PRS) no longer implies obtaining the data. In other words, the user has
to *opt into* getting the actual member data (by default, we retrieve only
the metadata for all members).

We introduce an abstraction of PRX+PRD/PRS. Specifically, we provide a new
struct and API centered around "PresageArchives" in the abstract (although
the source code subdirectory and the executable tool are still called "prx"
for now).

The prx(1) tool now uses PresageArchive under the hood. If the user provides
2 filenames to the tool instead of 1, then the tool assumes you've provided
a PRD and a PRS separately. If the user provides only 1 filename, then it's
assumed to be a single PRX, just like before.

The Lua bindings for PRX are upgraded to actually be bindings for
PresageArchive.

- Fixes #14
- Fixes #51