XPK: Support "nested repeat" feature
It turns out that having a single slot to keep track of the "repeat" status
while decoding XPKs is insufficient. There are lots of examples of a
repeat-spec embedded within another repeat-spec. So we introduce a stack of
repeat instructions.

Lots more XPKs are now fully decodable with this change.