CLU: Get rid of CLU_NewFromData() after all.
The intention behind the *_NewFromData() initiative is so that we can obtain
a structure (say, a CLU*) from either a standalone file and/or the archive
it's contained within. Turns out there's one API which can handle both
cases -- specifically one where the caller provides a FILE* on purpose. This
FILE* might be at the start of the file (in the case of a standalone CLU) or
somewhere in the middle (in the case of a PRX member).

The Lua binding to CLU is also updated.