Add xpkview(1), a SDL2-based XPK image viewer.
xpkview(1) is a simple graphical program written against SDL2. Right now it
has only the most basic of functionality.

It accepts a CLU and a XPK file on the command line, and it will render the
XPK image to screen. There is a faint checkerboard pattern in the background
to indicate transparency. You can use the LEFT and RIGHT arrow keys to view
the different frames of those XPKs which happen to have multiple frames in
them.